cf-sample-app-python/requirements.txt
Flask>=0.12
pyjwt>=1.7.1
cryptography>=2.9.2